sending and receiving the correct position data across network
may need refining, need to test with alex

+++ IndespectusProject/Assets/Scripts/NetworkingScripts/RoomScript.cs
@@ -150,7 +150,7 @@ public class RoomScript : MonoBehaviourPunCallbacks, IInRoomCallbacks {
             
             
             if(readyToCount || readyToStart) {
-                countdownText.SetText("Starting in: " + timeToStart);
+                countdownText.SetText("Starting in: " + (int)timeToStart);
             } else {countdownText.SetText("Game startting in: n/a");}
             if(!isGameLoaded) {
                 if(readyToStart) {
